Some major changes.
Fixed a bug in the graph. There were unused code and the Graph-function didn't do anything if the edge in the path were missing. Might have to refactor this function even further.

Try to get the scenarios, answerkeys and explanations into their own file, such that the templates can be reused. Just swap the graph to get new questions.

Need to add the questions over again and make new ones which can be verified by Job. Try to cover the whole asthma guideline.