Solved the issue temporally with warning at the charts.
We now have a code smell as API-calls shouldn't be done in the component. Need to rewrite the action meant for this purpose, but it needs to be written in such a way that we don't get loop-calls to redux. Loop calls to redux is what causes warnings and errors.